@@ -1,14 +1,14 @@
 name: flare_checkbox
 description: Checkbox widget with two or three state animation with a Flare animation of your choice 
-version: 1.0.4
+version: 2.0.0
 homepage: https://github.com/jaumard/flare_checkbox.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.7.8 <2.0.0"
 
 dependencies:
-  flare_flutter: ^2.0.1
+  flare_flutter: ^3.0.0
 
   flutter:
     sdk: flutter
